Changing backlight file
I made modification in order to use M4 to set the brightness and
max_brightness files. The used files are the one writen in the
config.mk file. To select the file of the system, just change this
paths.

TODO: install.sh file that asks for the files (or even better, that
finds the files)